# internal/dto/auth.go
package dto

// LoginRequest adalah DTO untuk request login
// @Description Request untuk login user
type LoginRequest struct {
	Email    string `json:"email" binding:"required,email" example:"user@example.com"`
	Password string `json:"password" binding:"required" example:"password123"`
}

// RegisterRequest adalah DTO untuk request register
// @Description Request untuk register user baru
type RegisterRequest struct {
	Email    string `json:"email" binding:"required,email" example:"user@example.com"`
	Password string `json:"password" binding:"required,min=6" example:"password123"`
	Name     string `json:"name" binding:"required" example:"John Doe"`
}

// LoginResponse adalah DTO untuk response login
// @Description Response setelah login berhasil
type LoginResponse struct {
	Token string       `json:"token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
	User  UserResponse `json:"user"`
}

// UserResponse dipindahkan ke dto/user.go untuk konsistensi